feat: Pause & Cancel Downloads
